backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pymupdf
python-dotenv
ebooklib
beautifulsoup4